requirements.txt
flask
flask-cors
python-dotenv
google
google.genai
google-generativeai
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pymongo